% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spineClustering.R
\name{plotBIC}
\alias{plotBIC}
\title{Plot BIC values}
\usage{
plotBIC(model)
}
\arguments{
\item{model}{a Mclust object}
}
\value{
None
}
\description{
For a given model plots the score obtained according to BIC criteria for each number of clusters
}
\examples{
plotBIC(model)

}
