% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spineClustering.R
\name{spineClustering}
\alias{spineClustering}
\title{Clustering dendritic spines by their morphology}
\usage{
spineClustering(csvSpines, numClusters = c(2:15), scale = T)
}
\arguments{
\item{csvSpines}{string value which indicates the path to the csv file where are saved the morphological features of the spines previously generated with the 3DSpineMFE library}

\item{scale}{boolean value that indicates if the data should be standardized}

\item{G}{a vector of integers with the number of clusters to try}
}
\value{
mclust object that contains the mixture of multivariate Gaussians
}
\description{
Read a csv file with the features of the morphology of the dendritic spines and cluster dendritic spines according to a mixture model of multivariate Gaussians.
}
\examples{
model <- spineClustering(csvSpines = system.file("extdata", "data.csv", package = "SpineMC"), numClusters = c(2:15), scale = T)

}
