\name{SpineMC-package}

\alias{SpineMC}

\title{3D morphological clustering and simulation of human dendritic spines}
\description{
Dendritic spines are small membranous protusions. They are receptors of excitatory synapses placed on the surface of some neuronal
dendrites. Their morphology have captured the attention of neuroscientific community because it has been associated with brain funcionality. For example, it has been claimed that thin spines contribute to learning, while the biggest and steady spines are linked to the memory process. Disturbances of their morphology or density have been related to mental disorders such as schizophrenia, dementia or mental retardation. Thus, the clustering of dendritic spines is an interest approach to better understand the previously describe relation between morphology and functionality. The present R package allows clustering of 3D dendritic spines according to their morphology.
}

\details{
Package: SpineMC

Type: Package

Version: 1.0

License: GPL(>=2)
}

\author{
Sergio Luengo-Sanchez
}
